changed how tillering method is set
