+++ Models/PMF/Structure/LeafCulms.cs
@@ -30,27 +30,21 @@ namespace Models.PMF.Struct
 	[PresenterName("UserInterface.Presenters.PropertyPresenter")]
 	public class LeafCulms : Model
 	{
-		/// <summary>
-		/// Link to the plant model.
-		/// </summary>
+		/// <summary>Link to the plant model.</summary>
 		[Link]
 		private Plant plant = null;
 
-		/// <summary>
-		/// Link to clock (used for FTN calculations at time of sowing).
-		/// </summary>
+		/// <summary>The parent leaf class</summary>
 		[Link]
-		private IClock clock = null;
+		SorghumLeaf leaf = null;
 
-		/// <summary>
-		/// Link to weather.
-		/// </summary>
-		[Link]
-		private IWeather weather = null;
-
-		/// <summary> Tillering Method to use for calculating how many tillers </summary>
+		/// <summary> Tillering Method that uses a fixed number of tillers</summary>
 		[Link(Type = LinkType.Child, ByName = true)]
-		private ITilleringMethod tillering = null;
+		private ITilleringMethod fixedTillering = null;
+
+		/// <summary> Tillering Method that manages number of tillers dynamically</summary>
+		[Link(Type = LinkType.Child, ByName = true)]
+		private ITilleringMethod dynamicTillering = null;
 
 		/// <summary>
 		/// Expansion stress.
@@ -140,13 +134,7 @@ namespace Models.PMF.Struct
 		[Link(Type = LinkType.Child, ByName = true)]
 		private IFunction numberOfLeaves = null;
 
-		/// <summary>
-		/// If true, tillering will be calculated on the fly.
-		/// Otherwise, number of tillers must be supplied via fertile
-		/// tiller number in the sowing rule.
-		/// </summary>
-		[Description("Dynamic tillering enabled")]
-		public bool DynamicTillering { get; set; }
+		private ITilleringMethod tillering => leaf.TilleringMethod == 0 ? fixedTillering : dynamicTillering;
 
 		/// <summary>
 		/// 
@@ -278,8 +266,6 @@ namespace Models.PMF.Struct
 		private void OnSowing(object sender, EventArgs e)
 		{
 			culmParams.Density = plant.SowingData.Population;
-			if (sender == this.plant && plant.SowingData.BudNumber == -1 && !DynamicTillering)
-				plant.SowingData.BudNumber = CalculateFtn();
 		}
 
 		///// <summary>
@@ -331,7 +317,6 @@ namespace Models.PMF.Struct
 				dltStressedLAI = dltPotentialLAI * expStress;
 				Culms.ForEach(c=> c.DltStressedLAI = c.DltLAI * expStress);
 			}
-
 		}
 
 		/// <summary>
@@ -344,166 +329,5 @@ namespace Models.PMF.Struct
 
 			return actualLAI;
         }
-
-		private double CalculateFtn()
-		{
-			// Estimate tillering given latitude, density, time of planting and
-			// row configuration. this will be replaced with dynamic
-			// calculations in the near future. Above latitude -25 is CQ, -25
-			// to -29 is SQ, below is NNSW.
-			double intercept, slope;
-
-			if (weather.Latitude > -12.5 || weather.Latitude < -38.0)
-				// Unknown region.
-				throw new Exception("Unable to estimate number of tillers at latitude {weather.Latitude}");
-
-			if(weather.Latitude > -25.0)
-			{
-				// Central Queensland.
-				if (clock.Today.DayOfYear < 319 && clock.Today.DayOfYear > 182)
-				{
-					// Between 1 July and 15 November.
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double (2.0).
-						intercept = 0.5786; slope = -0.0521;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5).
-						intercept = 0.8786; slope = -0.0696;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 1.1786; slope = -0.0871;
-					}
-				}
-				else
-				{
-					// After 15 November.
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double (2.0).
-						intercept = 0.4786; slope = -0.0421;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5)
-						intercept = 0.6393; slope = -0.0486;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 0.8000; slope = -0.0550;
-					}
-				}
-			}
-			else if(weather.Latitude > -29.0)
-			{
-				// South Queensland.
-				if (clock.Today.DayOfYear < 319 && clock.Today.DayOfYear > 182)
-				{
-					// Between 1 July and 15 November. 
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double  (2.0).
-						intercept = 1.1571; slope = -0.1043;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5).
-						intercept = 1.7571; slope = -0.1393;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 2.3571; slope = -0.1743;
-					}
-				}
-				else
-				{
-					// After 15 November.
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double (2.0).
-						intercept = 0.6786; slope = -0.0621;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5).
-						intercept = 1.1679; slope = -0.0957;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 1.6571; slope = -0.1293;
-					}
-				}
-			}
-			else
-			{
-				// Northern NSW.
-				if(clock.Today.DayOfYear < 319 && clock.Today.DayOfYear > 182)
-				{
-					//  Between 1 July and 15 November. 
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double (2.0).
-						intercept = 1.3571; slope = -0.1243;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5).
-						intercept = 2.2357; slope = -0.1814;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 3.1143; slope = -0.2386;
-					}
-				}
-				else if (clock.Today.DayOfYear > 349 || clock.Today.DayOfYear < 182)
-				{
-					// Between 15 December and 1 July.
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double (2.0).
-						intercept = 0.4000; slope = -0.0400;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5).
-						intercept = 1.0571; slope = -0.0943;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 1.7143; slope = -0.1486;
-					}
-				}
-				else
-				{
-					// Between 15 November and 15 December.
-					if (plant.SowingData.SkipRow > 1.9)
-					{
-						// Double (2.0).
-						intercept = 0.8786; slope = -0.0821;
-					}
-					else if (plant.SowingData.SkipRow > 1.4)
-					{
-						// Single (1.5).
-						intercept = 1.6464; slope = -0.1379;
-					}
-					else
-					{
-						// Solid (1.0).
-						intercept = 2.4143; slope = -0.1936;
-					}
-				}
-			}
-
-   			return Math.Max(slope * plant.SowingData.Population + intercept, 0);
-		}
 	}
 }
